requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
protobuf==3.20.3